Current status:
Can add points to the Model and have it reflected in both the Table and the Plot. The plot displays metadata about the points on hover. The delete buttons in the table can be used to remove points, which are reflected in the Model, the Table and the Plot. PlotZoom/TableSort are maintained throughout updates.

todo: unit and substance changes. Add auxiliary lines to plot. prettify. Reduce jank.